backend/requirements.txt
Django==6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1

# Production server
gunicorn==21.2.0

# PostgreSQL support
psycopg2-binary==2.9.9

# For development (optional in production)
django-debug-toolbar==4.3.0
